refactor: rename aws folder
